Make rdesc load dependent on resource data load in EDIT action
TinyMCE inputs were failing to bind to an input model when the
EDIT action was loaded without a cached model. This was because
they were instantiating their child scopes before the actual
model data was created, thus hiding the original model from the
parent scope. So it is necessary for us to load the actual
resource data BEFORE we load the resource description, so that
we don't try to generate the form (and bind to model) before it
has been populated with real data.

Resolves #95